+++ starkbank/workspace/__workspace.py
@@ -10,6 +10,7 @@ class Workspace(Resource):
     ## Parameters (required):
     - username [string]: Simplified name to define the workspace URL. This name must be unique across all Stark Bank Workspaces. Ex: "starkbankworkspace"
     - name [string]: Full name that identifies the Workspace. This name will appear when people access the Workspace on our platform, for example. Ex: "Stark Bank Workspace"
+    ## Parameters (optional):
     - allowed_tax_ids [list of strings]: list of tax IDs that will be allowed to send Deposits to this Workspace. ex: ["012.345.678-90", "20.018.183/0001-80"]
     ## Attributes:
     - id [string, default None]: unique id returned when the workspace is created. ex: "5656565656565656"